Moved Alerts to respond with a Boolean Property instead of an Optional<ButtonType> This change allows us to use non blocking alert types.
Have refactored three of the four places that used it.
Still need to refactor client update as that is very complex.

Currently works well apart from that and style changes.

Note that this alert will only do true or false, it does not support
some of the other alert types, although we do not use them (currently).

#fix #story[1054]