rework text alerts (alerts where you enter text - this is only used for setting/editing the override reason)
They now pass in both the instructions and pre-populate the textfield (if you are editing an existing override reason).
Also rearranged touch version so it looks nice (similar to normal touch alerts).
Added a getRawOverrideReason() to DonatedOrgan to get the bit of text that the user edits (ie not the time and user).